Create feeds view model
